requirements.txt
google-generativeai>=0.8.3
openai>=1.84.0
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
lxml==5.3.0
pyyaml>=6.0.1
mypy>=1.16.0
types-requests>=2.32.0
types-PyYAML>=6.0.0 